Display per-level-count at the top when browsing.
Note that the count for the previous level is also displayed in the breadcrumb when drilling down to the next level.

I’m uncertain as to whether this matters, but if it’s not wanted, the count could be displayed off to the side instead of making it part of the title string.